+++ src/main/java/com/marwit23/cook/dish/DishServiceImpl.java
@@ -1,6 +1,8 @@
 package com.marwit23.cook.dish;
 
 import com.marwit23.cook._exception.EntityNotFoundException;
+import com.marwit23.cook.todo.ToDoDish;
+import com.marwit23.cook.todo.ToDoDishRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -15,10 +17,14 @@ import java.util.Optional;
 public class DishServiceImpl implements DishService {
 
     private DishRepository dishRepository;
+    private DishIngredientRepository dishIngredientRepository;
+    private ToDoDishRepository toDoDishRepository;
 
     @Autowired
-    public DishServiceImpl(DishRepository dishRepository) {
+    public DishServiceImpl(DishRepository dishRepository, DishIngredientRepository dishIngredientRepository, ToDoDishRepository toDoDishRepository) {
         this.dishRepository = dishRepository;
+        this.dishIngredientRepository = dishIngredientRepository;
+        this.toDoDishRepository = toDoDishRepository;
     }
 
     @Override
@@ -49,6 +55,18 @@ public class DishServiceImpl implements DishService {
 
     @Override
     public void deleteById(Long dishId) {
-        dishRepository.deleteById(dishId);
+        Optional <Dish> result = dishRepository.findById(dishId);
+        Dish theDish = new Dish();
+        if(result.isPresent()) theDish = result.get();
+            for(DishIngredient dishIngredient : theDish.getDishIngredients()){
+                dishIngredient.setDish(null);
+                dishIngredient.setIngredient(null);
+                dishIngredientRepository.delete(dishIngredient);
+            }
+            for(ToDoDish toDoDish : theDish.getToDoDishList()){
+                toDoDish.setDish(null);
+                toDoDishRepository.delete(toDoDish);
+            }
+            dishRepository.deleteById(dishId);
     }
 }
